Refactor template content generators into helper class
